BECCS Liquid Fuels Assets
Creation of assets to transform biomass into liquid fuels. 1. Bio Gasoline, and 2. Bio FT with a mix of Gasoline, Jet Fuel, and Diesel. The transformation file is similar to BECCS Electricity and Hydrogen, but with fuel production rates, and the categorization of emissions into venting and biofuels emissions. The user have to ensure that carbon content have to be balanced such that biomass carbon = venting emissions + biofuel emissions + captured. The end node of the biofuels production goes directly into the respective fuels' end use demand node, thus the emissions of utilizing the biofuel is accounted in the transformation of BECCS_Liquid_Fuels and BECCS_Gasoline assets. This also allows the user to specify their own biofuel emission rates for biofuels produced from the specific process. Just ensure that carbon have to be balanced so the biomass carbon content either have to be emitted (vent + biofuels) and captured (if modeling process with CCS),