CO2 Storage Cap Constraint and Multiple CO2 Injection + Storage in each zone
Developed a constraint to limit the amount of CO2 stored in a co2 storage node with commodity CO2Captured. This is similar to the CO2 cap policy applied on the co2 sink node with commodity CO2. Now we can model CO2 storage injection costs and rates using the asset co2 injection, that leads to a co2 storage node that we can apply the CO2 storage cap constraint. This is like modeling geological storage with a max capacity per year. The user can also model multiple CO2 injection and storage sites in the same zone that connects to the same captured CO2 node, that represen the captured CO2 blending of the zone. This is shown in the example system where SE has 11 CO2 injection and storage sites. The typical pathway of captured CO2 is shown: 1. Capture from BECCS/DAC/CCS -> 2. Captured CO2 node in zone z -> 3. CO2 injection -> 4. CO2 storage node with CO2 storage cap constraint. Alternatively, the captured CO2 can be utilized by synthetic fuel resources instead of going to CO2 injection.