New example system without fuels, update to bio liquid fuel asset
Now there is two example systems, 1. eastern_us_three_zones_fuels with complete CO2, NG and liquid fuels supply chains, that include their exogeneous demand, include modeling of synthetic and biofuels, upstream and end use emissions. 2. eastern_us_three_zones_no_fuels that is almost the same as original example system with only electricity and H2 demand, and NG is a simple price input, but in this example system the CO2 supply chain is modeled with actual CO2 injection and storage instead of a sink.

A slight modification to bio liquid fuel production asset is that now it can also produce electricity in addition to fuels.